throw error for invalid dates
